+++ src/refactoring/components/Pages/Card/CartPage.tsx
@@ -1,7 +1,7 @@
-import { CartItem, Coupon, Product } from "../../../../types";
+import {  Coupon, Product } from "../../../../types";
 import { useCart } from "../../../hooks";
 import { getAppliedDiscount } from "../../../hooks/utils/cartUtils";
-import CartDetails from "./Units/CartDetails/CartDetatils";
+import ProductList from "./Units/ProductList/ProductList";
 
 interface Props {
   products: Product[];
@@ -25,7 +25,7 @@ export const CartPage = ({ products, coupons }: Props) => {
     <div className="container mx-auto p-4">
       <h1 className="text-3xl font-bold mb-6">장바구니</h1>
       <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-        <CartDetails
+        <ProductList
         products={products}
         cart={cart}
         addToCart={addToCart}
